@@ -104,20 +104,20 @@
     "rollup-plugin-commonjs": "^10.1.0",
     "rollup-plugin-sourcemaps": "^0.6.3",
     "rollup-plugin-typescript2": "^0.34.1",
-    "rxjs": "6.6.7",
+    "rxjs": "^7.8.1",
     "semantic-release": "19.0.5",
     "ts-jest": "^26.5.6",
     "tslint": "^6.1.3",
     "tslint-config-prettier": "^1.18.0",
     "typedoc": "^0.22.17",
-    "typescript": "^4.0.5"
+    "typescript": "^4.2"
   },
   "dependencies": {
     "lodash-es": "^4.17.21"
   },
   "peerDependencies": {
     "core-js": "^3.6.5",
-    "rxjs": "^6.5.5"
+    "rxjs": "^7.0.0"
   },
   "repository": {
     "type": "git",
